+++ src/FluxFlow.Components.Storage/Timing/IStorageClock.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.Storage.Timing;
+
+public interface IStorageClock
+{
+    DateTimeOffset UtcNow { get; }
+}